fix Mac tests / update build agent
* (test_osn_input) add missing hotkeys for MAC_SCREEN_CAPTURE
* (nodeobs_settings.cpp) fix "&& within ||" Xcode warning

(test_nodeobs_settings) do not set ApplyServiceSettings to true on mac

* If this property is enabled on Mac, then the encoder settings will be overwritten
and cause the test to fail- because you're not allowed to change certain fields.

rename macos_avcapture to fix test

* avcapture plugins were renamed so they can be looked up directly in JS code need to update the tests- https://github.com/streamlabs/obs-studio/pull/679/commits/b53043d5bb8072ad9a07ae1109d2f988ed84fdc6

(Video.cpp - CreateVideo) add bounds check on input from JS

(test_osn_input) ignore ndi_source (exists on client but not on agent)
(deleteConfigFiles) check if folder exists first before trying to remove files
* fix issue where it would stop deleting config files on the mac-Crashpad folder.
now all config files are deleted properly and the ini files will be properly
reset before next test

(main.yml) update CI steps to pass along build artifacts
(controller.cpp) print out error if unable to spawn obs64 (helpful on CI build agent)

+++ obs-studio-client/source/controller.cpp
@@ -313,9 +313,14 @@ std::shared_ptr<ipc::client> Controller::host(const std::string &uri)
 	remove(uri.c_str());
 
 	int ret = posix_spawnp(&pid, serverBinaryPath.c_str(), NULL, NULL, argv, environ);
+	if (ret != 0) {
+		std::cerr << "Could not spawn the server at " << serverBinaryPath.c_str() << " with error code: " << ret << std::endl;
+		return nullptr;
+	}
 	// Connect
 	std::shared_ptr<ipc::client> cl = connect(uri);
 	if (!cl) { // Assume the server broke or was not allowed to run.
+		std::cerr << "Could not connect to ipc::server" << std::endl;
 		disconnect();
 		uint32_t exitcode;
 		kill(pid, SIGKILL);
